Introduce BaseGraphWriterDAO test base class
